reorganize app initialization
