+++ packages/fxa-settings/src/lib/config.test.ts
@@ -13,13 +13,13 @@ import config, {
 beforeEach(reset);
 
 describe('readConfigMeta', () => {
-  // it('throws an error when the meta tag element cannot be found', () => {
-  //   expect(() =>
-  //     readConfigMeta(() => {
-  //       return null;
-  //     })
-  //   ).toThrowError('<meta name="fxa-config"> is missing');
-  // });
+  it('throws an error when the meta tag element cannot be found', () => {
+    expect(() =>
+      readConfigMeta(() => {
+        return null;
+      })
+    ).toThrowError('<meta name="fxa-config"> is missing');
+  });
 
   it('throws an error when the content of the meta tag is not decodable', () => {
     expect(() =>
