+++ packages/fxa-settings/src/models/index.ts
@@ -1,3 +1,4 @@
-export * from './App';
+export * from './AppContext';
 export * from './Account';
 export * from './Session';
+export * from './hooks';
